ENH: Add GetDefaultParameterMapGTest GoogleTest
Moved the ParameterObject tests that are specifically testing `ParameterObject::GetDefaultParameterMap` to a separate cxx file, and a separate GoogleTest "test suite".
